Substituição e acréscimo de cores
Sabendo que na maioria dos hardware as operações com int's são mais otimizados, então foi substituído todos os unsigned short para ter a melhor desempenho no programa, foi colocado cores no X e no OK do resultado de Avaliação

+++ Problema_Operadores_bit_a_bit/binario.cpp
@@ -1,4 +1,4 @@
-int ligar_bit(unsigned short* valor_teste, int numero_bit) {
+int ligar_bit(int* valor_teste, int numero_bit) {
 
 	int resultado_ligar, mascara = 1;
 	resultado_ligar = *valor_teste | (mascara << numero_bit);
@@ -6,7 +6,7 @@ int ligar_bit(unsigned short* valor_teste, int numero_bit) {
 	return resultado_ligar;
 
 }
-int desliga_bit(unsigned short* valor_teste, int numero_bit) {
+int desliga_bit(int* valor_teste, int numero_bit) {
 
 	int mascara = 1, desliga_resultado;
 
@@ -15,7 +15,7 @@ int desliga_bit(unsigned short* valor_teste, int numero_bit) {
 	return desliga_resultado;
 }
 
-bool testa_bit(unsigned short* valor_teste, int numero_bit) {
+bool testa_bit(int* valor_teste, int numero_bit) {
 	
 	int mascara = 1;
 	if (*valor_teste & (mascara << numero_bit)) {
@@ -27,18 +27,18 @@ bool testa_bit(unsigned short* valor_teste, int numero_bit) {
 	}
 }
 
-int binario_and(unsigned short* valor_teste_1, unsigned short* valor_teste_2) {
+int binario_and(int* valor_teste_1, int* valor_teste_2) {
 
 	return *valor_teste_1 & *valor_teste_2;
 
 }
-int binario_or(unsigned short* valor_teste_1, unsigned short* valor_teste_2) {
+int binario_or(int* valor_teste_1, int* valor_teste_2) {
 	return *valor_teste_1 | *valor_teste_2;
 }
-int bits_baixos(unsigned short* valor_teste,int i) {
+int bits_baixos(int* valor_teste,int i) {
 
 	return valor_teste[i] & 255;
 }
-int bits_altos(unsigned short* valor_teste,int i) {
+int bits_altos(int* valor_teste,int i) {
 	return valor_teste[i] & 65280;
 }
