+++ Problema_Operadores_bit_a_bit/binario.h
@@ -1,8 +1,8 @@
 #pragma once
-int ligar_bit(unsigned short* testea, int numero_bit);
-int desliga_bit(unsigned short* testea, int numero_bit);
-bool testa_bit(unsigned short* valor_teste, int numero_bit);
-int binario_and(unsigned short* valor_teste_1, unsigned short* valor_teste_2);
-int binario_or(unsigned short* valor_teste_1, unsigned short* valor_teste_2);
-int bits_baixos(unsigned short* valor_teste,int i);
-int bits_altos(unsigned short* valor_teste, int i);
+int ligar_bit(int* testea, int numero_bit);
+int desliga_bit(int* testea, int numero_bit);
+bool testa_bit(int* valor_teste, int numero_bit);
+int binario_and(int* valor_teste_1, int* valor_teste_2);
+int binario_or(int* valor_teste_1, int* valor_teste_2);
+int bits_baixos(int* valor_teste,int i);
+int bits_altos(int* valor_teste, int i);
